move: 엔티티 패키지 변경 (레이어우선 -> 모듈우선 구조)
