+++ common/src/main/java/com/food/common/common/domain/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.food.common.domain;
+package com.food.common.common.domain;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
