+++ common/src/main/java/com/food/common/foodCategory/repository/FoodCategoryRepository.java
@@ -1,6 +1,6 @@
-package com.food.common.repository;
+package com.food.common.foodCategory.repository;
 
-import com.food.common.domain.FoodCategory;
+import com.food.common.foodCategory.domain.FoodCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FoodCategoryRepository extends JpaRepository<FoodCategory, Long> {
