+++ common/src/main/java/com/food/common/payment/domain/Payment.java
@@ -1,6 +1,8 @@
-package com.food.common.domain;
+package com.food.common.payment.domain;
 
-import com.food.common.vo.Point;
+import com.food.common.common.domain.BaseTimeEntity;
+import com.food.common.user.domain.User;
+import com.food.common.common.domain.Point;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
